--story=856902347 【All】支持服务扩展 - 以流水线构件列表页的构件操作为例，完成一个功能为通过作业平台部署构件的小扩展

+++ src/backend/ci/ext/tencent/store/biz-store-service/src/main/kotlin/com/tencent/devops/store/pojo/ExtServiceVersionLogCreateInfo.kt
@@ -6,7 +6,7 @@ data class ExtServiceVersionLogCreateInfo(
     @ApiModelProperty("扩展服务ID")
     val serviceId: String,
     @ApiModelProperty("发布类型，0：新上架 1：非兼容性升级 2：兼容性功能更新 3：兼容性问题修正  ")
-    val releaseType: Int,
+    val releaseType: Byte,
     @ApiModelProperty("版本日志内容")
     val content: String,
     @ApiModelProperty("添加用户")
